requirements.txt
django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
django-celery-beat==2.9.0
celery==5.6.3
redis==7.4.0
psycopg2-binary==2.9.12
gunicorn==23.0.0
dj-database-url==2.3.0
python-dotenv==1.0.1
